apollo-cli/requirements.txt
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
tenacity>=8.2.0
